Move relationship_analysis to models directory
